move home components into home folder
