+++ src/pages/Home/AppStatusCard.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
-import AppIcon from "../../../container/AppIcon/AppIcon";
-import { AppStatus } from "../../../models/app-status";
-import { availableApps } from "../../../util/availableApps";
+import AppIcon from "../../container/AppIcon/AppIcon";
+import { AppStatus } from "../../models/app-status";
+import { availableApps } from "../../util/availableApps";
 
 type Props = {
   app: AppStatus;
