+++ src/pages/Home/HardwareCard.tsx
@@ -1,7 +1,7 @@
 import { FC, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { SSEContext } from "../../../store/sse-context";
-import LoadingBox from "../../Shared/LoadingBox/LoadingBox";
+import { SSEContext } from "../../store/sse-context";
+import LoadingBox from "../../components/Shared/LoadingBox/LoadingBox";
 
 const PI_NUM_CORES = 4;
