+++ src/pages/Home/LightningCard.tsx
@@ -1,10 +1,10 @@
 import { FC, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { AppContext, Unit } from "../../../store/app-context";
-import { SSEContext } from "../../../store/sse-context";
-import { convertMSatToBtc, convertToString } from "../../../util/format";
-import { checkPropsUndefined } from "../../../util/util";
-import LoadingBox from "../../Shared/LoadingBox/LoadingBox";
+import LoadingBox from "../../components/Shared/LoadingBox/LoadingBox";
+import { AppContext, Unit } from "../../store/app-context";
+import { SSEContext } from "../../store/sse-context";
+import { convertMSatToBtc, convertToString } from "../../util/format";
+import { checkPropsUndefined } from "../../util/util";
 
 export const LightningCard: FC = () => {
   const { t } = useTranslation();
