+++ src/pages/Home/TransactionCard/TransactionDetailModal/LNDetails/LNDetails.tsx
